+++ tests/unit_tests/service/document/test_update_document_service.py
@@ -16,23 +16,6 @@ def _to_write_dto(dto: DocumentReadDTO) -> DocumentWriteDTO:
     )
 
 
-def test_document_service_get_upload_details(test_s3_client):
-    result = doc_service.get_upload_details("path/file.ext", True)
-    assert result is not None
-    assert len(result) == 2
-    # Check the signed url starts with the right path
-    assert result[0].startswith(
-        "https://test-document-bucket.s3.amazonaws.com/path/file.ext?"
-    )
-    assert "X-Amz-Algorithm" in result[0]
-    assert "X-Amz-Credential" in result[0]
-    assert "X-Amz-Date" in result[0]
-    assert "X-Amz-Expires=3600" in result[0]
-    assert "X-Amz-Signature" in result[0]
-
-    assert result[1] == "https://cdn.climatepolicyradar.org/path/file.ext"
-
-
 # --- UPDATE
